server/requirements.txt
flask==3.0.3
Pillow==10.4.0
datasets==2.20.0
matplotlib==3.9.1
pandas==2.2.2
xgboost==2.1.0
scikit-learn==1.5.1
seaborn==0.13.2
joblib==1.4.2